move workspace_command::clean to separate module
